+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog (Alpha Versions)
 ## [Release Alpha 1.4.2](https://github.com/Nevah5/wordlebot/releases/tag/a1.4.2) - 09.02.2022
+### Added
+- Selection menu for stats in new module `./src/modules/interactions.js`
+
 ### Fixed
 - Slash command bug on /stat when giving no user